Add financial and sponsorship information
I structed the information here in a way that people could zoom in or
link to different docs or sections as they need to know answers to different
aspects of how we collect and handle money. This is a wip and will be
changed with the input of others. Also, some parts are still blank, but
I wanted to put out the initial structure and get feedback.