Remove cash advances, direct pay updates
We don't have a policy in place for cash advances yet, so we're removing this from the documentation. The goal is still to figure this piece out but we want to keep things simple for now.  And, we can do direct pay for independent communities as well so I added that detail.